Move routes global into agent
